Updates for photon original direction capture
For polarization it is necessary to store the original photon direction. The photon which travels to the observer is a 'virtual' photon and it's original scattering direction is gone, hence this commit stores the l m and n.
Also the binning has been devreased here..this is mainly for the analytical test and needs to be restored in the final commit to dev.